requirements.txt
Jinja2==3.1.6
pandas==2.2
openpyxl==3.1.5
python-dotenv==1.0.1